# openapi.yaml
openapi: 3.0.1
info:
  title: Website Scraping Plugin
  description: A plugin that allows the user to find hotels that meet search criteria using ChatGPT. If you do not know the user's username, ask them first before making queries to the plugin. Otherwise, use the username "global".
  version: 'v1'
paths:
  # /todos/{username}:
  #   get:
  #     operationId: getTodos
  #     summary: Get the list of todos
  #     parameters:
  #     - in: path
  #       name: username
  #       schema:
  #           type: string
  #       required: true
  #       description: The name of the user.
  #     responses:
  #       "200":
  #         description: OK
  #         content:
  #           application/json:
  #             schema:
  #               $ref: '#/components/schemas/getTodosResponse'
  #   post:
  #     operationId: addTodo
  #     summary: Add a todo to the list
  #     parameters:
  #     - in: path
  #       name: username
  #       schema:
  #           type: string
  #       required: true
  #       description: The name of the user.
  #     requestBody:
  #       required: true
  #       content:
  #         application/json:
  #           schema:
  #             $ref: '#/components/schemas/addTodoRequest'
  #     responses:
  #       "200":
  #         description: OK
  #   delete:
  #     operationId: deleteTodo
  #     summary: Delete a todo from the list
  #     parameters:
  #     - in: path
  #       name: username
  #       schema:
  #           type: string
  #       required: true
  #       description: The name of the user.
  #     requestBody:
  #       required: true
  #       content:
  #         application/json:
  #           schema:
  #             $ref: '#/components/schemas/deleteTodoRequest'
  #     responses:
  #       "200":
  #         description: OK
  /website/{username}:
    post:
      operationId: getWebsite
      summary: Fetch the website content at a url
      parameters:
      - in: path
        name: username
        schema:
            type: string
        required: true
        description: The name of the user.
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/getWebsiteRequest'
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/getWebsiteResponse'
  
  /hotel/{username}:
    post:
      operationId: getHotelAvailability
      summary: Get hotel availability
      parameters:
      - in: path
        name: username
        schema:
            type: string
        required: true
        description: The name of the user.
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/getHotelRequest'
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/getHotelResponse'

components:
  schemas:
    # getTodosResponse:
    #   type: object
    #   properties:
    #     todos:
    #       type: array
    #       items:
    #         type: string
    #       description: The list of todos.
    # addTodoRequest:
    #   type: object
    #   required:
    #   - todo
    #   properties:
    #     todo:
    #       type: string
    #       description: The todo to add to the list.
    #       required: true
    # deleteTodoRequest:
    #   type: object
    #   required:
    #   - todo_idx
    #   properties:
    #     todo_idx:
    #       type: integer
    #       description: The index of the todo to delete.
    #       required: true
    getWebsiteRequest:
       type: object
       required:
       - url
       properties:
         url:
           type: string
           description: The url of the website to fetch
           required: true  
    getWebsiteResponse:
      type: object
      properties:
        summarized_html:
          type: string
          description: The summarized website content at a url
          required: true
        actions:
          type: array
          items:
            type: object
            properties:
              url: 
                type: string
                description: the url of the available action
              action_description:
                type: string
                description: The action available at that URL
    getHotelRequest:
       type: object
       required:
       - name
       - startDate
       - endDate
       properties:
         name:
           type: string
           description: The name of the hotel
           required: true
         startDate:
           type: string
           description: The starting date of the reservation in the format MM/DD/YYYY
           required: true
         endDate:
           type: string
           description: The ending date of the reservation in the format MM/DD/YYYY
           required: true
         numRooms:
           type: string
           description: The number of rooms to book. Assume 1 if not specified.
           required: false
         numAdults:
           type: string
           description: The number of adults in the booking. Assume 2 if not specified.
           required: false
         numChildren:
           type: string
           description: The number of children in the booking. Assume 0 if not specified.
           required: false
           
    getHotelResponse:
      type: object
      properties:
        availability:
          type: string
          description: "The javascript code used to make the reservation on the website"
          required: true
           